requirements.txt
# Core ML libraries
torch>=2.0.0
transformers>=4.30.0
diffusers>=0.21.0
accelerate>=0.20.0

# Image processing
Pillow>=9.5.0
opencv-python>=4.8.0

# MLflow for experiment tracking
mlflow>=2.5.0

# Streamlit for demo UI
streamlit>=1.25.0

# Data science essentials
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Jupyter notebooks
jupyter>=1.0.0
ipykernel>=6.23.0

# Testing
pytest>=7.4.0

# Utilities
tqdm>=4.65.0
requests>=2.31.0 